pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tcg-pregrader"
version = "0.1.0"
description = "CNN-based TCG card pre-grading system (Pokemon MVP)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "typer>=0.12.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "tensorflow>=2.16.0",
    "opencv-python>=4.9.0",
    "pillow>=10.3.0",
    "structlog>=24.2.0",
]

[project.optional-dependencies]
dev = [
    "hypothesis>=6.100.0",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
# CLI entry point — `pregrader predict ...` resolves to cli.app
pregrader = "pregrader.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
# Hypothesis CI profile is loaded in conftest.py

[tool.coverage.run]
source = ["src/pregrader"]
omit = ["tests/*"]
